deploy(NNC): update from nn_compress
